refactor: tests and tests runner

+++ src/lib/clock-native.c
@@ -7,5 +7,5 @@
 Value clockNative(int argCount, Value* args) {
   arityCheck(0, argCount);
   
-  return NUMBER_VAL(clock() / CLOCKS_PER_SEC);
+  return NUMBER_VAL(clock() / (double) CLOCKS_PER_SEC);
 }
